split C and CPP interfaces
