+++ psi_cardinality/c/psi_cardinality_server_c.h
@@ -24,36 +24,36 @@
 extern "C" {
 #endif
 
-typedef void* psi_cardinality_server_ctx;
+typedef void *psi_cardinality_server_ctx;
 
 struct server_buffer_t {
-    const char* buff;
-    size_t buff_len;
+  const char *buff;
+  size_t buff_len;
 };
 
 psi_cardinality_server_ctx psi_cardinality_server_create_with_new_key();
-psi_cardinality_server_ctx psi_cardinality_server_create_from_key(
-    struct server_buffer_t key_bytes);
-void psi_cardinality_server_delete(psi_cardinality_server_ctx* ctx);
+psi_cardinality_server_ctx
+psi_cardinality_server_create_from_key(struct server_buffer_t key_bytes);
+void psi_cardinality_server_delete(psi_cardinality_server_ctx *ctx);
 
 int psi_cardinality_server_create_setup_message(psi_cardinality_server_ctx ctx,
                                                 double fpr,
                                                 int64_t num_client_inputs,
-                                                struct server_buffer_t* input,
-                                                size_t input_len, char** output,
-                                                size_t* output_len);
+                                                struct server_buffer_t *input,
+                                                size_t input_len, char **output,
+                                                size_t *output_len);
 void psi_cardinality_server_delete_buffer(psi_cardinality_server_ctx ctx,
-                                          char** input);
+                                          char **input);
 
 int psi_cardinality_server_process_request(
     psi_cardinality_server_ctx ctx, struct server_buffer_t client_request,
-    char** output, size_t* output_len);
+    char **output, size_t *output_len);
 int psi_cardinality_server_get_private_key_bytes(psi_cardinality_server_ctx ctx,
-                                                 char** output,
-                                                 size_t* output_len);
+                                                 char **output,
+                                                 size_t *output_len);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif  // PSI_CARDINALITY_PSI_CARDINALITY_SERVER_C_H_
+#endif // PSI_CARDINALITY_PSI_CARDINALITY_SERVER_C_H_
